Fix broken relative import path after file moving
